refactor(editor): place the note menu with the other surfaces
Move menu next to keymap and the mobile toolbar as its own lint owner, and drop the unused adapters allow.
