+++ config/eslint/boundaries.ts
@@ -14,7 +14,6 @@ const FEATURE_TYPES = [
   'features-date',
   'features-folding',
   'features-links',
-  'features-menu',
   'features-note-body',
   'features-search',
   'features-zoom',
@@ -47,13 +46,13 @@ export const editorBoundaries = {
     element('features-date', `${EDITOR}/features/date`),
     element('features-folding', `${EDITOR}/features/folding`),
     element('features-links', `${EDITOR}/features/links`),
-    element('features-menu', `${EDITOR}/features/menu`),
     element('features-note-body', `${EDITOR}/features/note-body`),
     element('features-search', `${EDITOR}/features/search`),
     element('features-zoom', `${EDITOR}/features/zoom`),
     element('outline', `${EDITOR}/outline`),
     element('runtime', `${EDITOR}/runtime`),
     element('keymap', `${EDITOR}/keymap`),
+    element('menu', `${EDITOR}/menu`),
     element('mobile-toolbar', `${EDITOR}/mobile-toolbar`),
     element('triggers', `${EDITOR}/triggers`),
     element('view', `${EDITOR}/view`),
@@ -86,8 +85,8 @@ export const editorBoundaries = {
     // Workspace re-exports the pending-import API for the app.
     { from: { element: { type: 'view' } },
       allow: { to: { element: { type: ['outline', 'features-search', 'features-zoom', 'runtime'] } } } },
-    { from: { element: { type: 'features-menu' } },
-      allow: { to: { element: { type: ['foundation', 'outline', 'runtime', 'triggers', 'view', 'adapters', 'features-list-types', 'features-folding'] } } } },
+    { from: { element: { type: 'menu' } },
+      allow: { to: { element: { type: ['foundation', 'outline', 'runtime', 'triggers', 'view', 'features-list-types', 'features-folding'] } } } },
     { from: { element: { type: FEATURE_TYPES } },
       allow: { to: { element: { type: ['foundation', 'outline', 'runtime', 'triggers', 'view', 'adapters'] } } } },
     // The shell composes, so it reaches everything it mounts. Its edge into
@@ -100,6 +99,7 @@ export const editorBoundaries = {
         'runtime',
         'outline',
         'keymap',
+        'menu',
         'mobile-toolbar',
         'editor-dev',
         ...EDITING_TYPES,
