+++ src/client/editor/menu/NoteMenuPlugin.tsx
@@ -18,9 +18,9 @@ import { requireContentItemFromNode } from '#client/editor/outline/schema';
 import { installOutlineSelectionHelpers } from '#client/editor/outline/selection/store';
 import { $canOfferFold } from '#client/editor/features/folding/fold-offer';
 import { $getNestedListType } from '#client/editor/features/list-types/nested-list-type';
-import { handleNoteMenuShortcut } from '#client/editor/features/menu/note-menu-shortcuts';
-import type { NoteMenuShortcutEvent } from '#client/editor/features/menu/note-menu-shortcuts';
-import { $resolveNoteStateFromDOMNode } from '#client/editor/features/menu/note-state';
+import { handleNoteMenuShortcut } from '#client/editor/menu/note-menu-shortcuts';
+import type { NoteMenuShortcutEvent } from '#client/editor/menu/note-menu-shortcuts';
+import { $resolveNoteStateFromDOMNode } from '#client/editor/menu/note-state';
 import { isOtherPopupActive, setPopupActive } from '#client/editor/triggers/active-popup';
 import { EditorPopupOverlay } from '#client/editor/triggers/overlay';
 import { resolveCaretTargetRect } from '#client/editor/triggers/target-rect';
