+++ src/client/editor/menu/note-menu-shortcuts.ts
