+++ src/client/editor/menu/note-state.ts
