+++ src/client/editor/shell/Editor.tsx
@@ -25,8 +25,8 @@ import { StatusIndicators } from './StatusIndicators';
 import { ZoomPlugin } from '#client/editor/features/zoom/ZoomPlugin';
 import { ZoomVisibilityPlugin } from '#client/editor/features/zoom/ZoomVisibilityPlugin';
 import { FoldingPlugin } from '#client/editor/features/folding/FoldingPlugin';
-import { NoteControlsPlugin } from '#client/editor/features/menu/NoteControlsPlugin';
-import { NoteMenuPlugin } from '#client/editor/features/menu/NoteMenuPlugin';
+import { NoteControlsPlugin } from '#client/editor/menu/NoteControlsPlugin';
+import { NoteMenuPlugin } from '#client/editor/menu/NoteMenuPlugin';
 import { MobileActionToolbar } from '#client/editor/mobile-toolbar/MobileActionToolbar';
 import { SearchCandidatesPlugin } from '#client/editor/features/search/SearchCandidatesPlugin';
 import { PendingDocumentImportPlugin } from '#client/editor/runtime/PendingDocumentImportPlugin';
